Dk cleanup (#442)
* separate ftwalkcommon.c from dk functions

* renaming dk files, start working on attribute names

* more attribute and other DK naming / labeling
