+++ include/functions.h
@@ -1302,7 +1302,7 @@ void func_8007925C();
 // ? func_8007A06C(?);
 void func_8007AB48();
 void func_8007AB80();
-void func_8007ABD0(struct _Hitbox*, u32, struct _HSD_GObj*);
+void func_8007ABD0(struct _Hitbox*, u32 damageAmount, struct _HSD_GObj*);
 // ? func_8007AC68(?);
 // ? func_8007AC9C(?);
 // ? func_8007AD18(?);
@@ -1782,8 +1782,8 @@ void func_8008A2BC(struct _HSD_GObj* gobj);
 // ? func_8008D3A4(?);
 // ? func_8008D708(?);
 f32 func_8008D7F0(struct _Fighter*);
-// ? func_8008D8D8(?);
-// ? func_8008D8E8(?);
+f32 scaleBy154_8008D8D8(f32);  
+s32 func_8008D8E8(f32);
 void func_8008D930();
 // ? func_8008DA4C(?);
 // ? func_8008DB10(?);
@@ -2670,12 +2670,12 @@ void func_800DF0D0();
 // ? ftDonkey_800DFA70(?);
 // ? ftDonkey_800DFA98(?);
 // ? ftDonkey_800DFAE4(?);
-// ? ftDonkey_GetWalkType_800DFBF8(?);
-// ? ftDonkey_800DFC70(?);
-// ? ftDonkey_800DFCA4(?);
-// ? ftDonkey_800DFDDC(?);
-// ? ftDonkey_800DFEC8(?);
-// ? ftDonkey_800E0060(?);
+// ? ftWalkCommon_GetWalkType_800DFBF8(?);
+// ? ftWalkCommon_800DFC70(?);
+// ? ftWalkCommon_800DFCA4(?);
+// ? ftWalkCommon_800DFDDC(?);
+// ? ftWalkCommon_800DFEC8(?);
+// ? ftWalkCommon_800E0060(?);
 // ? ftDonkey_800E0134(?);
 // ? ftDonkey_800E017C(?);
 // ? ftDonkey_800E0294(?);
