+++ obj_files.mk
@@ -76,8 +76,8 @@ TEXT_O_FILES := 						            \
 	$(BUILD_DIR)/asm/melee/ft/ftlib.s.o 		\
 	$(BUILD_DIR)/asm/melee/ft/code_800877F8.s.o 		\
 	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey1.c.o 	\
-	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey2.c.o 	\
-	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey3.c.o 	\
+	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey_walk.c.o 	\
+	$(BUILD_DIR)/src/melee/ft/ftwalkcommon.c.o 	\
 	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey4.c.o 	\
 	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey5.c.o 	\
 	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey6.c.o 	\
@@ -106,11 +106,11 @@ TEXT_O_FILES := 						            \
 	$(BUILD_DIR)/asm/melee/ft/chara/ftlink.s.o 		\
 	$(BUILD_DIR)/src/melee/ft/chara/ftKirby/ftkirby.c.o 		\
 	$(BUILD_DIR)/asm/melee/ft/chara/ftkirby.s.o 		\
-	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey10.c.o 	\
-	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey11.c.o 	\
-	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey12.c.o 	\
-	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey13.c.o 	\
-	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey14.c.o 	\
+	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey_callbacks.c.o 	\
+	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey_SpecialLw.c.o 	\
+	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey_SpecialS.c.o 	\
+	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey_SpecialN.c.o 	\
+	$(BUILD_DIR)/src/melee/ft/chara/ftDonkey/ftdonkey_SpecialHi.c.o 	\
 	$(BUILD_DIR)/asm/melee/ft/chara/ftdonkey_2.s.o 	\
 	$(BUILD_DIR)/src/melee/ft/chara/ftSeak/ftseak.c.o 		\
 	$(BUILD_DIR)/asm/melee/ft/chara/ftseak.s.o 		\
