+++ src/melee/ft/chara/ftDonkey/ftdonkey.h
@@ -1,38 +1,63 @@
 #pragma once
 
 #include "melee/ft/ftcommon.h"
+#include "melee/ft/ftwalkcommon.h"
 #include "melee/lb/lbvector.h"
 
 typedef struct _ftDonkeyAttributes {
-    s32 x0;
-    f32 x4;
+    s32 action_state;
+    s32 x4_action_state;
+
+    //// used by ftdonkey_walk.c (maybe Cargo Hold) ////
     f32 x8;
     f32 xC;
     f32 x10;
     f32 x14;
     f32 x18;
     f32 x1C;
-    f32 x20;
-    f32 x24;
-    f32 x28;
-    s32 x2C;
-    s32 x30;
-    f32 x34;
-    f32 x38;
-    f32 x3C;
-    f32 x40;
-    f32 x44;
-    f32 x48;
-    f32 x4C;
-    f32 x50;
-    f32 x54;
-    f32 x58;
-    f32 x5C;
-    f32 x60;
-    f32 x64;
-    f32 x68;
-    f32 x6C;
-    f32 x70;
+    /////////////////////////////////////////////////
+
+    struct {
+        f32 x20_TURN_SPEED;
+        f32 x24_JUMP_STARTUP_LAG;
+        f32 x28_LANDING_LAG;
+    } cargo_hold;
+
+    //// Giant Punch
+    struct {
+        s32 x2C_MAX_ARM_SWINGS;
+        s32 x30_DAMAGE_PER_SWING;
+        f32 x34_PUNCH_HORIZONTAL_VEL;
+        f32 x38_LANDING_LAG;
+    } SpecialN;
+
+    //// Headbutt
+    struct {
+        f32 x3C_MIN_STICK_X_MOMENTUM;
+        f32 x40_MOMENTUM_TRANSITION_MODIFIER;
+        f32 x44_AERIAL_GRAVITY;
+    } SpecialS;
+
+    f32 x48_UNKNOWN;
+
+    //// Spinning Kong
+    struct {
+        f32 x4C_AERIAL_VERTICAL_VELOCITY;
+        f32 x50_AERIAL_GRAVITY;
+        f32 x54_GROUNDED_HORIZONTAL_VELOCITY;
+        f32 x58_AERIAL_HORIZONTAL_VELOCITY;
+        f32 x5C_GROUNDED_MOBILITY;
+        f32 x60_AERIAL_MOBILITY;
+        f32 x64_LANDING_LAG;
+    } SpecialHi;
+
+    /// Hand Slap
+    struct {
+        f32 x68;
+        f32 x6C;
+        f32 x70;
+    } SpecialLw;
+
 } ftDonkeyAttributes;
 
 void ftDonkey_800DF980(HSD_GObj* fighterObj);
@@ -41,14 +66,7 @@ s32 ftDonkey_800DFA98(HSD_GObj* fighterObj);
 
 void ftDonkey_800DFAE4(HSD_GObj* fighterObj, f32);
 void ftDonkey_800DFB54(HSD_GObj* fighterObj);
-void ftDonkey_800DFCA4(HSD_GObj *fighterObj, s32 arg1, s32 arg2, f32 arg8, f32 arg9, f32 argA, f32 argB, f32 argC, f32 argD, f32 argE, f32 argF);
-void ftDonkey_800DFDDC(HSD_GObj* fighterObj);
-void ftDonkey_800DFEC8(HSD_GObj* fighterObj, void (*arg_cb)(HSD_GObj*, f32));
 
-
-void ftDonkey_800DFDDC(HSD_GObj* fighterObj);
-
-void ftDonkey_800E0060(HSD_GObj* fighterObj);
 void ftDonkey_800E017C(HSD_GObj* fighterObj);
 void ftDonkey_800E0294(HSD_GObj* fighterObj);
 void ftDonkey_800E05E4(HSD_GObj* fighterObj);
