+++ src/melee/ft/chara/ftDonkey/ftdonkey1.c
@@ -16,5 +16,20 @@ void ftDonkey_800DF980(HSD_GObj* fighterObj) {
         func_8007D7FC(fighter);
     }
     donkey_attr = getFtSpecialAttrs2CC(fighter);
-    Fighter_ActionStateChange_800693AC(fighterObj, donkey_attr->x0, 0, NULL, 0.0f, 1.0f, 0.0f);
+    Fighter_ActionStateChange_800693AC(fighterObj, donkey_attr->action_state, 0, NULL, 0.0f, 1.0f, 0.0f);
+}
+
+void ftDonkey_800DF9F0(HSD_GObj* fighterObj) {
+    if ((!func_80094EA4(fighterObj)) && (!ftDonkey_800E0378(fighterObj)) && (!ftDonkey_800E0134(fighterObj))) {
+        s32 result = ftDonkey_800DFA98(fighterObj);
+        if (result) { return;  }
+    }
+}
+
+void ftDonkey_800DFA50(HSD_GObj* fighterObj) {
+    func_80084F3C(fighterObj);
+}
+
+void ftDonkey_800DFA70(HSD_GObj* fighterObj) {
+    func_8008403C(fighterObj, &ftDonkey_800E0294);
 }
