+++ src/melee/ft/chara/ftDonkey/ftdonkey7.c
@@ -19,10 +19,10 @@ void ftDonkey_800E04A4(HSD_GObj* fighterObj, s32 arg1) {
     ftDonkeyAttributes* donkey_attr2;
     fighter->x2344_stateVar2_s32 = arg1;
     fighter->x2340_stateVar1 = 0;
-    fighter->x2348_stateVar3_f32 = donkey_attr->x24;
+    fighter->x2348_stateVar3_f32 = donkey_attr->cargo_hold.x24_JUMP_STARTUP_LAG;
     donkey_attr2 = getFtSpecialAttrs2CC(fighter);
     
-    Fighter_ActionStateChange_800693AC(fighterObj, donkey_attr2->x0 + 5, 0, NULL, 0.0f, 1.0f, 0.0f);
+    Fighter_ActionStateChange_800693AC(fighterObj, donkey_attr2->action_state + 5, 0, NULL, 0.0f, 1.0f, 0.0f);
     ftAnim_SetAnimRate(fighterObj, 0.0f);
 }
